+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0 - 2025-03-15
+
+- Freeze animations with `F`
+- Toggle background color with `G`
+
 ## 0.1.0 - 2025-03-15
 
 - Initial Release